Move FieldViewHolder to its own class
